# [exporter/signalfx] Translate metrics together instead of individually