… exporter/signalfxexporter/testdata/json/hostmetrics.json