**Description:** <Describe what has changed.>

Translation rules were originally done on metrics individually. This means that only datapoints within the same metric could be operated on together. This is insufficient for the new host metrics format removing the direction dimension. To properly form metrics to match their previous format, translations must be done across metrics, especially aggregations. This only changes the fact that multiple MTS will be operated on at the same time, it doesn't change any translation rules. The solution is to do aggregations at the `ScopeMetric` metric level, so metrics within the same `ScopeMetric` can be aggregated.

This change also properly changes test metric data to be their proper format, moving metrics into corresponding `ResourceMetrics` and `ScopeMetrics`, instead of all being lumped into the same slices.

This is an internal refactoring and testing change that should not impact any input or output data. Because of this, I don't think there should be any CHANGELOG entry.

**Link to tracking Issue:** <Issue number if applicable>

This PR is the first step to solving issue #12641

**Testing:** <Describe what testing was performed and which tests were added.>

Ensured existing tests pass, data being emitted is the same as before change.

**Documentation:** <Describe the documentation added.>

NA since this is a refactor without any external impact.