@crobert-1 2022-08-12
> Where is the actual change now? :)

The change at this point is just updating test data format. It could go with another PR that will update translation rules, but I figure it would be good to have the format change attached to the discussion we've had on this PR.