@dmitryax 2022-08-12
but we still apply `translateAndFilter` on data points from one `pdata.Metric`, the idea is to apply it on set of data points from one `pdata.ScopeMetrics` object