@crobert-1 2022-08-12
> but we still apply `translateAndFilter` on data points from one `pdata.Metric`, the idea is to apply it on set of data points from one `pdata.ScopeMetrics` object

I don't know why, but I got mixed around with this review. You're right, we want to apply it at the `ScopeMetrics` level. For some reason I got confused and thought the inner-most loop was the scope level, rather than individual metric level.

I'll update the description again and move translateAndFilter call.