@crobert-1 2022-08-19
I've added a benchmark test that takes hostmetrics and times how long it takes to run `MetricsToSignalFxV2`. The only change to this functionality was the scope of the translation loop. 

Here's the performance comparison before and after this change on the same dataset:
Before:
```
BenchmarkMetricConversion-12       	    2332	    504435 ns/op
BenchmarkMetricConversion-12       	    2166	    558897 ns/op
BenchmarkMetricConversion-12       	    1760	    637230 ns/op
```

After:
```
BenchmarkMetricConversion-12       	    2235	    488777 ns/op
BenchmarkMetricConversion-12       	    2451	    488565 ns/op
BenchmarkMetricConversion-12       	    2424	    514079 ns/op
```
I believe this shows there is a slight performance improvement for the signalfx exporter with the new functionality.