@dmitryax 2022-08-19
@crobert-1 thanks for adding the benchmark 👍  Can you please post before and after results?