@dmitryax 2022-08-19
Nice! Seems like there is a clear improvement. I believe we can merge it regardless of `duration` attribute's changes